requirements.txt
Flask==2.3.2
flask-socketio==5.3.6
eventlet==0.33.3
numpy==1.25.1
sounddevice==0.4.7
librosa==0.10.0
openai-whisper==20230314
torch==2.0.1
transformers==4.30.2
sentencepiece==0.1.99
pydub==0.25.1
streamlit==1.32.0
streamlit-audiorecorder==0.0.3
scipy==1.11.3
speechbrain==0.5.16
python-dotenv==1.0.0
Werkzeug==2.3.7 